Use `systemd-resolved` for name resolution
This commit changes name resolution to use much-hated systemd-resoved.
The main reason for this (besides annoying systemd-haters :-) is that
it makes DNS configuration much easier for images that run in multiple
environments.

A fallback DNS is provided too for environments where DNS is not configured
(or supplied by DHCP) - currently using Quad9 and Cloudflare public DNS.