Install `qemu-guest-agent` by default
This commit adds `qemu-guest-agent` to list of "basic" packages. Quite
likely the image will be run under KVM and so QEMU GA would be useful.

When not, QEMU-GA service won't be started at all (it is started
conditionally when QEMU/KVM virtualization is detected) and space-wise,
it is some ~200k so no bother.