Update `grub2.sh` to support bare metal images
This commit updates `grub2.sh` to allow installing GRUB into
images that are intended to run wide range of real hardware.

Prior this commit, `grub2.sh`-installed boot was optimized for
QEMU/KVM virtual machines (serial console, minimal initramfs
and so on).

With this commit by using new option `-g` the boot process is
made a generic as possible, allowing one to use `dd` to install
image onto real hardware and boot it.